Delete, Update employee
Adding delete emp method to UI and it works fine
Adding edit emp method to UI and it works fine with dynamically reflecting from the emp info from the url by passing the emp id in the url.
Now we are done with the technical part. Lets move to the styling part...